implementação crud completa

+++ lib/commands/students/subcommands/delete_by_id_command.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:args/command_runner.dart';
 import 'package:lab_aluno_cli/repositories/student_repository.dart';
@@ -26,7 +27,14 @@ class DeleteByIdCommand extends Command {
       return;
     }
     final id = int.parse(argResults?['id']);
-    await repository.deleteById(id);
-    print('deletado com sucesso');
+    final student = await repository.findById(id);
+    print('deseja realmente deletar o (${student.id}) - ${student.name}? (s ou n)');
+    final confirm = stdin.readLineSync();
+    if (confirm!.toLowerCase() == 's') {
+      await repository.deleteById(id);
+      print('deletado com sucesso');
+      return;
+    }
+    print('operação cancelada');
   }
 }
